ACTUAL remaining implicit casts removed

+++ src/local.hpp
@@ -153,7 +153,7 @@ struct linedef_t
 
   uint32_t flags;
   uint32_t special;
-  int32_t tag;
+  uint16_t tag;
 
   bool dont_render;
   bool dont_render_front;
